[CHORE/#15] 컴포저블 파일화
