+++ app/src/main/java/com/sopt/bubble/feature/precise_store/component/PreciseStoreTicket.kt
@@ -1,0 +1,77 @@
+package com.sopt.bubble.feature.precise_store.component
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import com.sopt.bubble.R
+import com.sopt.bubble.ui.theme.Body03
+import com.sopt.bubble.ui.theme.Gray500
+import com.sopt.bubble.ui.theme.Gray800
+import com.sopt.bubble.ui.theme.Name02
+import com.sopt.bubble.ui.theme.Name03
+import com.sopt.bubble.ui.theme.White
+
+@Composable
+fun PreciseStoreTicket(
+    title: String,
+    price: String,
+    modifier: Modifier = Modifier,
+    originalPrice: String? = null,
+) {
+    Card(
+        shape = RoundedCornerShape(
+            topStart = 10.dp, topEnd = 10.dp,
+            bottomStart = 10.dp, bottomEnd = 10.dp
+        ),
+        colors = CardDefaults.cardColors(containerColor = Gray800),
+        modifier = modifier
+            .fillMaxWidth()
+            .padding(horizontal = 20.dp)
+    ) {
+        Row(
+            verticalAlignment = Alignment.CenterVertically,
+            horizontalArrangement = Arrangement.SpaceBetween,
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(vertical = 22.5.dp, horizontal = 12.dp)
+        ) {
+            Text(
+                text = stringResource(id = R.string.precise_store_ticket_number, title),
+                color = White,
+                style = Name03
+            )
+            Row(
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                if (originalPrice != null) {
+                    Text(
+                        text = originalPrice,
+                        color = Gray500,
+                        style = Body03,
+                    )
+                }
+                Text(
+                    text = stringResource(id = R.string.precise_store_ticket_price, price),
+                    color = White,
+                    style = Name02,
+                    modifier = Modifier.padding(horizontal = 5.dp)
+                )
+                Text(
+                    text = stringResource(id = R.string.precise_store_ticket_month),
+                    color = Gray500,
+                    style = Body03
+                )
+            }
+        }
+    }
+}
