+++ app/src/main/java/com/sopt/bubble/feature/precise_store/model/CheckBoxContent.kt
@@ -13,6 +13,8 @@ data class CheckBoxContent(
     var onClickCheckBox:() -> Unit = {}
 )
 
+const val CHECK_BUTTON_NUM = 3
+
 val checkBoxList: List<CheckBoxContent> = listOf(
     CheckBoxContent(
         title = R.string.precise_store_checkbox_title01,
